+++ src/plugins/value.js
@@ -1,4 +1,4 @@
-import { getValueAt, setValueAt as _setValueAt } from '../utils';
+import { getValueAt, setValueAt } from '../utils';
 
 export default function(context, stateProvider) {
 
@@ -19,7 +19,7 @@ export default function(context, stateProvider) {
         if (arguments.length < 1) {
             return getValueAt(data, columns, position);
         }
-        _setValueAt(data, columns, rows, position, value);
+        setValueAt(data, columns, rows, position, value);
         return context;
     }
 
@@ -39,38 +39,9 @@ export default function(context, stateProvider) {
         if (arguments.length < 2) {
             return getValueAt(data, columns, pos);
         }
-        _setValueAt(data, columns, rows, pos, value);
+        setValueAt(data, columns, rows, pos, value);
         return context;
     }
 
-    /**
-     * Set the value at the current position. You can also set the cell to <code>undefined</code>
-     * The current position can be defined by [goto(position)]{@link gridl#goto} or [walk(direction)]{@link gridl#walk}.
-     *
-     * @deprecated Use [value()]{@link gridl#value} instead. Will be removed in version >= 0.9.x
-     *
-     * @memberOf gridl
-     * @method
-     * @instance
-     *
-     * @param {*} value - The value the cell should have.
-     */
-    const setValue = value => valueAt(position, value);
-
-    /**
-     * Set the value at a certain position. You can also set the cell to <code>undefined</code>
-     *
-     * @deprecated Use [valueAt()]{@link gridl#valueAt} instead. Will be removed in version > 0.9.x
-     *
-     * @memberOf gridl
-     * @method
-     * @instance
-     *
-     * @param {Array.<number>} pos - The position where you want to set the value.
-     * @param {*} value - The value you want to set.
-     * @returns {gridl} The same gridl instance.
-     */
-    const setValueAt = (pos, value) => valueAt(pos, value);
-
-    return { value, valueAt, setValue, setValueAt };
+    return { value, valueAt };
 }