+++ src/plugins/value.js
@@ -3,26 +3,6 @@ import { getValueAt, setValueAt } from '../utils';
 export default function(context, state) {
 
     /**
-     * Get or set the value at the current position.<br>
-     * It returns the cell's value if you provide no value and sets it if you do provide a value.
-     *
-     * @memberOf gridl
-     * @method
-     * @instance
-     *
-     * @param {*} [value] - The value you want to set or <code>undefined</code> if you want to get the value.
-     * @returns {*} The cell's value or the gridl instance if you use it as a setter.
-     */
-    function value(value) {
-        const { data, columns, rows, position } = state;
-        if (arguments.length < 1) {
-            return getValueAt(data, columns, position);
-        }
-        setValueAt(data, columns, rows, position, value);
-        return context;
-    }
-
-    /**
      * Get or set the value at a certain position.<br>
      * It returns the cell's value if you provide no value and sets it if you do provide a value.
      *
@@ -43,5 +23,5 @@ export default function(context, state) {
         return context;
     }
 
-    return { methods: { value, valueAt } };
+    return { methods: { valueAt } };
 }