+++ docs/changelog.md
@@ -4,8 +4,8 @@
 
 * Plugin API !!!
     * You can now add custom functions to gridl and mess around with the internal state. Note: all responsibility lies with the plugin creator!
-    * Heavy internal refactoring (every public gridl function is now a plugin by definition)
-    * support for namespaces
+    * heavy internal refactoring (every public gridl implemented via a plugin)
+    * support for namespaces: functions can be scoped with the plugins' name
 * removed `setValue()` and `setValueAt()`, now fully replaced by `value()` and `valueAt()`
 * renamed mirror functions
     * `mirrorX()` to `flipX()`